Fix i2c and add color sensor driver code

+++ drivers/color.h
@@ -10,10 +10,10 @@ typedef struct _rgba_sensor_info * rgba_sensor_info;
 // structure representing rgb color values
 // each value is an 8-bit unsigned integer value.
 typedef struct rgba_col {
-    u8 r; // red
-    u8 g; // green
-    u8 b; // blue
-    u8 a; // alpha
+    float r; // red
+    float g; // green
+    float b; // blue
+    float a; // alpha
 } rgba_col;
 
 // initialize an rgb sensor with its relevant state information
